[refactor:move] desktop-runtime: tests/mod.rs のヘルパを tests/support/ へ分割
1,955行の tests/mod.rs から共有ヘルパ68アイテム(fn 65 + struct 3)を
用途別の support/{env,media,sync_status,social,replication,mock_cn,net}.rs
へ移動。mod.rs は import block + `mod support;` + glob 再輸出シム +
サブモジュール宣言のみの63行に縮小。

既存テスト13ファイル(55テスト)は無変更。純粋な行移動からの唯一の
逸脱は、移動アイテムと Mock struct フィールドへの pub(crate) 付与
(親は子の private を glob import できないため。crate 外公開は不変)。

検証: nextest list 84テスト完全一致、rust-test + e2e-smoke green、
clippy -D warnings / fmt green、oversized-files 解消(最大606行)。

Refs: .claude/plans/2026-07-03-refactor-s1-test-split.md (T2 / 付録B)
