+++ crates/desktop-runtime/src/tests/support/mod.rs
@@ -1,0 +1,14 @@
+mod env;
+mod media;
+mod mock_cn;
+mod net;
+mod replication;
+mod social;
+mod sync_status;
+pub(crate) use env::*;
+pub(crate) use media::*;
+pub(crate) use mock_cn::*;
+pub(crate) use net::*;
+pub(crate) use replication::*;
+pub(crate) use social::*;
+pub(crate) use sync_status::*;
